requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.15.0
huggingface_hub>=0.19.0
python-dotenv>=1.0.0